HS-120 FIX: $? 뒤에 바로 문자열이 들어올 경우 처리
